* windows compilation

+++ src/hotspot/share/gc/g1/g1UncommitRegionTask.cpp
@@ -106,7 +106,7 @@ void G1UncommitRegionTask::execute() {
 
   // Translate the size limit into a number of regions, taking the minimum
   // region count into account.
-  static const uint max_num_regions_to_uncommit = MAX2<uint>(UncommitSizeLimit / G1HeapRegionSize, MinRegionsToUncommit);
+  static const uint max_num_regions_to_uncommit = MAX2<uint>(UncommitSizeLimit / (uint)G1HeapRegionSize, MinRegionsToUncommit);
 
   // Prevent from running during a GC pause.
   SuspendibleThreadSetJoiner sts;
